Apply all SolveSpace relevant patches except f67996f6 and 8f958955
4fdd4f29	Not needed
Delete all allocated items in destructors.

85d33d89	Already upstream - not needed
Read and write knot weights in DRW_Spline.

53e47a53	Applied
Add forgotten initialization for DRW_Dimension::length.

b11a2e86	Applied
Set $DIMSCALE default value to 1.0.

The old value was likely a copy-paste error.

7ce0a76e	Implemented differently upstream - not needed? hmmm...
Don't write $DIMTXSTY (handle of referenced STYLE) by default.

QCad doesn't display any dimensions where it's set.

680ad14b	Should I merge this?
Adjust DRW_Dimstyle defaults to match DRW_Header::write.

97f48718	Applied
Initialize DRW dimension type in dimension subclass constructors.

decd2369	Applied
Add DRW_Entity::setWidthMm.

In principle there's an algebraic relationship between the DXF line
width value and line thickness in millimeters, and everything except
AutoCAD accepts any value from the range, but AutoCAD ignores all
values except a few chosen ones.

48dd46d7	Applied
Don't use pure virtual functions in DRW_Interface.

This forced API consumers to implement a lot of useless empty
functions for no good reason.

a6f7c0a9	Not needed - fixed upstream
Mollify MSVC warnings.

ea19c78d	Applied
Add missing copy constructors.

d91f9fb7	Applied
Fix H/V align reading for MTEXT.

a05d16ea	Not applied
Remove unused DRW_Interface::setBlock.

e32b54b9	Not needed - fixed upstream
Silence a few -Wunused-variable warnings.

These are for a not-implemented features in DRW_Image, so it should
be OK.

8d24e656	Applied, `measureValue`, `hasmeasureValue` instead of `actual`, `hasActual`
Add DRW_Dimension::{has,get,set}ActualMeasurement.

e389be82	Not needed - fixed upstream

Don't add spline control points twice.

This caused a double free when deallocating DRW_Hatch.

145a36fb	Applied

Remove dwgR::testReader().

f67996f6	!!!!!!!!!!!!!! OFFFFFFFF !!!!!!!!!!!!!

Implement variants of read/write functions that take std::[io]stream*.

8f958955	!!!!!!!!!!!!!! OFFFFFFFF !!!!!!!!!!!!!

Remove the fileName field from dwgR/dxfRW.

It's more trouble than it's worth when supporting Windows. Instead,
require the caller to always use the appropriate overload of
std::[io]fstream constructor (possibly the wchar_t* one).

3475a9bf	Applied

In POLYLINE, use vertex type that matches the polyline type.

This makes polylines readable in AutoCAD again.

6f362317	Applied

Fix type conversion warnings.

2d45ee8c	Not needed - fixed upstream

Fix GCC warning -Wmisleading-indentation.

cbd3d7a4	Applied

Added more codes for colors supported by DXF.

4928b015	Applied

Add support for $TDCREATE.

87507fa7	Implemented diiferently upstrean - not needed.

Fix misplaced branch in DRW_TextCodec::setVersion

87ff1082	Applied

Fix opening files saved as DXF R10 in AutoCAD.

When saving in the AC1006 format, several DIMSTYLE groups were
erroneously omitted.

af191563	Applied

Up to DXF R12, the default layer "0" may be left undefined.

Some standards mandate the default layer to be unused.

d56caa42	Applied

Use DRW_Block::layer when writing blocks.

Before, the layer was hardcoded to "0" in the writer.

eca52c4f	Skipped to avoid unnecessary differences

Misc typos: length

Found via codespell and grep

f84dc2d1	Skipped to avoid unnecessary differences

Misc. source typos: compressed

Found via `codespell`

19d15979	Skipped to avoid unnecessary differences

Misc. typos

03fa5f30	Skipped to avoid unnecessary differences

Commit missing parts of eca52c4f.

ce3b0edc	Implemented differently upstream - not needed.

DRW_Coord: Simplify constructor/assignment.

Having a user-specified assignment operator but default copy-constructor
violates rule of 2/3/5/0.

Found by clazy.

5d2240fb	Already upstream - not needed

Add const-qualifiers to DRW_Dimension getters

getExtrusion and getName was missing const qualifiers.

0b7b7b70	Applied

Read extrusion tags for dimension entities

Dimension entities support DXF codes 210, 220, 230 and and DRW_Dimension
already had a extPoint member, but the tags were never used when
parsing.

+++ drw_objects.cpp
@@ -480,7 +480,7 @@ bool DRW_LType::parseCode(int code, const std::unique_ptr<dxfReader>& reader){
 /*TODO: control max length permited */
 void DRW_LType::update(){
     double d =0;
-    size = path.size();
+    size = (int)path.size();
     for (int i = 0;  i< size; i++){
         d += fabs(path.at(i));
     }
